backend/core/config.py
import os
from dotenv import load_dotenv

# Load .env at import time (EARLY)
load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

if not OPENAI_API_KEY:
    raise ValueError("OPENAI_API_KEY is not set in environment variables")
